requirements.txt
asyncqt==0.8.0
joblib==1.1.0
matplotlib==3.5.3
nest_asyncio==1.5.5
nfstream==6.5.2
numpy==1.23.2
pandas==1.4.3
psutil==5.9.2
PyQt5==5.15.7
pyshark==0.5.3
pytest==7.2.0
python-dotenv==0.21.0
scapy==2.4.5
scikit_learn==1.1.3
scipy==1.9.1
setuptools==63.2.0
winotify==1.1.0
